requirements.txt
playwright>=1.40.0
pyyaml>=6.0
pdfplumber>=0.10.0
anthropic>=0.30.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
twilio>=9.0.0
